# internal/platform/httpserver/openapi/api.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package openapi

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// Defines values for LivenessStatus.
const (
	Ok LivenessStatus = "ok"
)

// Valid indicates whether the value is a known member of the LivenessStatus enum.
func (e LivenessStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// Defines values for ReadinessStatus.
const (
	Degraded ReadinessStatus = "degraded"
	NotReady ReadinessStatus = "not_ready"
	Ready    ReadinessStatus = "ready"
)

// Valid indicates whether the value is a known member of the ReadinessStatus enum.
func (e ReadinessStatus) Valid() bool {
	switch e {
	case Degraded:
		return true
	case NotReady:
		return true
	case Ready:
		return true
	default:
		return false
	}
}

// Defines values for WhoamiRole.
const (
	ADMIN     WhoamiRole = "ADMIN"
	DEVELOPER WhoamiRole = "DEVELOPER"
	OPERATOR  WhoamiRole = "OPERATOR"
	SERVICE   WhoamiRole = "SERVICE"
	USER      WhoamiRole = "USER"
)

// Valid indicates whether the value is a known member of the WhoamiRole enum.
func (e WhoamiRole) Valid() bool {
	switch e {
	case ADMIN:
		return true
	case DEVELOPER:
		return true
	case OPERATOR:
		return true
	case SERVICE:
		return true
	case USER:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	// Code Stable machine-readable error code (UPPER_SNAKE_CASE).
	//
	// Example: UNAUTHENTICATED
	Code string `json:"code"`

	// Details Optional structured, client-safe detail.
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message Human-readable, client-safe message.
	//
	// Example: authentication required
	Message string `json:"message"`

	// RequestId Example: req_3f9c2a7d41b8e6a0
	RequestId string `json:"requestId"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"error"`
}

// Liveness defines model for Liveness.
type Liveness struct {
	Status LivenessStatus `json:"status"`
}

// LivenessStatus defines model for Liveness.Status.
type LivenessStatus string

// Readiness defines model for Readiness.
type Readiness struct {
	// Checks Result of each dependency check, `ok` or `failed`.
	Checks *map[string]string `json:"checks,omitempty"`
	Status ReadinessStatus    `json:"status"`
}

// ReadinessStatus defines model for Readiness.Status.
type ReadinessStatus string

// Whoami defines model for Whoami.
type Whoami struct {
	// ClientId Identifier of the API client.
	ClientId string     `json:"clientId"`
	Role     WhoamiRole `json:"role"`
}

// WhoamiRole defines model for Whoami.Role.
type WhoamiRole string

// Forbidden defines model for Forbidden.
type Forbidden = ErrorResponse

// Internal defines model for Internal.
type Internal = ErrorResponse

// RateLimited defines model for RateLimited.
type RateLimited = ErrorResponse

// Unauthenticated defines model for Unauthenticated.
type Unauthenticated = ErrorResponse

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetHealthz Liveness probe
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// GetReadyz Readiness probe
	// (GET /readyz)
	GetReadyz(w http.ResponseWriter, r *http.Request)
	// GetWhoami Identify the calling API client
	// (GET /v1/whoami)
	GetWhoami(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// GetHealthz Liveness probe
// (GET /healthz)
func (_ Unimplemented) GetHealthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetReadyz Readiness probe
// (GET /readyz)
func (_ Unimplemented) GetReadyz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetWhoami Identify the calling API client
// (GET /v1/whoami)
func (_ Unimplemented) GetWhoami(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReadyz operation middleware
func (siw *ServerInterfaceWrapper) GetReadyz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReadyz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWhoami operation middleware
func (siw *ServerInterfaceWrapper) GetWhoami(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWhoami(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealthz)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/readyz", wrapper.GetReadyz)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/whoami", wrapper.GetWhoami)
	})

	return r
}

type ForbiddenResponseHeaders struct {
	XRequestId *string
}
type ForbiddenJSONResponse struct {
	Body ErrorResponse

	Headers ForbiddenResponseHeaders
}

type InternalResponseHeaders struct {
	XRequestId *string
}
type InternalJSONResponse struct {
	Body ErrorResponse

	Headers InternalResponseHeaders
}

type RateLimitedResponseHeaders struct {
	RateLimitLimit     *int
	RateLimitRemaining *int
	RateLimitReset     *int
	RetryAfter         *int
	XRequestId         *string
}
type RateLimitedJSONResponse struct {
	Body ErrorResponse

	Headers RateLimitedResponseHeaders
}

type UnauthenticatedResponseHeaders struct {
	WWWAuthenticate *string
	XRequestId      *string
}
type UnauthenticatedJSONResponse struct {
	Body ErrorResponse

	Headers UnauthenticatedResponseHeaders
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse Liveness

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyzRequestObject struct {
}

type GetReadyzResponseObject interface {
	VisitGetReadyzResponse(w http.ResponseWriter) error
}

type GetReadyz200JSONResponse Readiness

func (response GetReadyz200JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyz503JSONResponse Readiness

func (response GetReadyz503JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetWhoamiRequestObject struct {
}

type GetWhoamiResponseObject interface {
	VisitGetWhoamiResponse(w http.ResponseWriter) error
}

type GetWhoami200ResponseHeaders struct {
	XRequestId *string
}

type GetWhoami200JSONResponse struct {
	Body    Whoami
	Headers GetWhoami200ResponseHeaders
}

func (response GetWhoami200JSONResponse) VisitGetWhoamiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.XRequestId != nil {
		w.Header().Set("X-Request-Id", fmt.Sprint(*response.Headers.XRequestId))
	}
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetWhoami401JSONResponse struct{ UnauthenticatedJSONResponse }

func (response GetWhoami401JSONResponse) VisitGetWhoamiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.WWWAuthenticate != nil {
		w.Header().Set("WWW-Authenticate", fmt.Sprint(*response.Headers.WWWAuthenticate))
	}
	if response.Headers.XRequestId != nil {
		w.Header().Set("X-Request-Id", fmt.Sprint(*response.Headers.XRequestId))
	}
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetWhoami403JSONResponse struct{ ForbiddenJSONResponse }

func (response GetWhoami403JSONResponse) VisitGetWhoamiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.XRequestId != nil {
		w.Header().Set("X-Request-Id", fmt.Sprint(*response.Headers.XRequestId))
	}
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type GetWhoami429JSONResponse struct{ RateLimitedJSONResponse }

func (response GetWhoami429JSONResponse) VisitGetWhoamiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.RateLimitLimit != nil {
		w.Header().Set("RateLimit-Limit", fmt.Sprint(*response.Headers.RateLimitLimit))
	}
	if response.Headers.RateLimitRemaining != nil {
		w.Header().Set("RateLimit-Remaining", fmt.Sprint(*response.Headers.RateLimitRemaining))
	}
	if response.Headers.RateLimitReset != nil {
		w.Header().Set("RateLimit-Reset", fmt.Sprint(*response.Headers.RateLimitReset))
	}
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	if response.Headers.XRequestId != nil {
		w.Header().Set("X-Request-Id", fmt.Sprint(*response.Headers.XRequestId))
	}
	w.WriteHeader(429)
	_, err := buf.WriteTo(w)
	return err
}

type GetWhoami500JSONResponse struct{ InternalJSONResponse }

func (response GetWhoami500JSONResponse) VisitGetWhoamiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	if response.Headers.XRequestId != nil {
		w.Header().Set("X-Request-Id", fmt.Sprint(*response.Headers.XRequestId))
	}
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetHealthz Liveness probe
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
	// GetReadyz Readiness probe
	// (GET /readyz)
	GetReadyz(ctx context.Context, request GetReadyzRequestObject) (GetReadyzResponseObject, error)
	// GetWhoami Identify the calling API client
	// (GET /v1/whoami)
	GetWhoami(ctx context.Context, request GetWhoamiRequestObject) (GetWhoamiResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(w http.ResponseWriter, r *http.Request) {
	var request GetHealthzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetReadyz operation middleware
func (sh *strictHandler) GetReadyz(w http.ResponseWriter, r *http.Request) {
	var request GetReadyzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetReadyz(ctx, request.(GetReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReadyz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetReadyzResponseObject); ok {
		if err := validResponse.VisitGetReadyzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetWhoami operation middleware
func (sh *strictHandler) GetWhoami(w http.ResponseWriter, r *http.Request) {
	var request GetWhoamiRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetWhoami(ctx, request.(GetWhoamiRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetWhoami")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetWhoamiResponseObject); ok {
		if err := validResponse.VisitGetWhoamiResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}